Rename :runtime-movile module -> :runtime-service
