--- library/runtime-mobile/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=runtime-mobile
-POM_NAME=KmpTor Runtime For Mobile
-POM_DESCRIPTION=Mobile specific runtime utilities
