+++ library/runtime-service/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=runtime-service
+POM_NAME=KmpTor Runtime For TorService
+POM_DESCRIPTION=TorRuntime.ServiceFactory specific runtime implementation
